Change move command steps params from int16 to int32.

+++ app/src/main/java/net/agolyakov/agoslider/service/bluetooth/BluetoothService.kt
@@ -259,7 +259,7 @@ class BluetoothService @Inject constructor(
         bleManager.writeHomeCommand(homeX, homeC, homeB)
     }
 
-    fun sendMoveCommand(x: Short, c: Short, b: Short) {
+    fun sendMoveCommand(x: Int, c: Int, b: Int) {
         bleManager.writeMoveCommand(x, c, b)
     }
